remove redunant donit component, linting, minor refactors
